work + fix linux build

+++ GAME/SAVEGAME.C
@@ -82,9 +82,9 @@ void RestoreLaraData(int FullSave)//538D0(<), 53D34(<) (F)
 	memcpy(&lara, &savegame.Lara, sizeof(struct lara_info));
 	lara.target = NULL;
 	lara.spaz_effect = NULL;
-	lara.right_arm.frame_base = (short*)((char*)lara.right_arm.frame_base + (unsigned int)objects[PISTOLS_ANIM].frame_base);
-	lara.left_arm.frame_base = (short*)((char *)lara.left_arm.frame_base + (unsigned int)objects[PISTOLS_ANIM].frame_base);
-	lara.GeneralPtr = (char *)lara.GeneralPtr + (unsigned int)malloc_buffer;
+	lara.right_arm.frame_base = (short*)((char*)lara.right_arm.frame_base + (ptrdiff_t)objects[PISTOLS_ANIM].frame_base);
+	lara.left_arm.frame_base = (short*)((char *)lara.left_arm.frame_base + (ptrdiff_t)objects[PISTOLS_ANIM].frame_base);
+	lara.GeneralPtr = (char *)lara.GeneralPtr + (ptrdiff_t)malloc_buffer;
 	if (lara.burn)
 	{
 		lara.burn = 0;
@@ -115,7 +115,7 @@ void RestoreLaraData(int FullSave)//538D0(<), 53D34(<) (F)
 	
 	for (i = 0; i < 15; i++)
 	{
-		lara.mesh_ptrs[i] = (short*)((char*)mesh_base + (int)lara.mesh_ptrs[i]);
+		lara.mesh_ptrs[i] = (short*)((char*)mesh_base + (ptrdiff_t)lara.mesh_ptrs[i]);
 	}
 
 	_CutSceneTriggered1 = savegame.CutSceneTriggered1;
@@ -129,12 +129,12 @@ void SaveLaraData()//53738(<), 53B9C(<) (F)
 
 	for (i = 0; i < 15; i++)
 	{
-		lara.mesh_ptrs[i] = (short*)((char*)lara.mesh_ptrs[i] - (int)mesh_base);
+		lara.mesh_ptrs[i] = (short*)((char*)lara.mesh_ptrs[i] - (ptrdiff_t)mesh_base);
 	}
 
-	lara.left_arm.frame_base = (short*)((char *)lara.left_arm.frame_base - (unsigned int)objects[PISTOLS_ANIM].frame_base);
-	lara.right_arm.frame_base = (short*)((char *)lara.right_arm.frame_base - (unsigned int)objects[PISTOLS_ANIM].frame_base);
-	lara.GeneralPtr = (char *)lara.GeneralPtr - (unsigned int)malloc_buffer;
+	lara.left_arm.frame_base = (short*)((char *)lara.left_arm.frame_base - (ptrdiff_t)objects[PISTOLS_ANIM].frame_base);
+	lara.right_arm.frame_base = (short*)((char *)lara.right_arm.frame_base - (ptrdiff_t)objects[PISTOLS_ANIM].frame_base);
+	lara.GeneralPtr = (char *)lara.GeneralPtr - (ptrdiff_t)malloc_buffer;
 	memcpy(&savegame.Lara, &lara, sizeof(savegame.Lara));
 	
 	for (i = 0; i < 15; i++)
@@ -142,9 +142,9 @@ void SaveLaraData()//53738(<), 53B9C(<) (F)
 		lara.mesh_ptrs[i] = (short*)((char*)mesh_base + (int)lara.mesh_ptrs[i]);
 	}
 
-	lara.left_arm.frame_base = (short*)((char *)lara.left_arm.frame_base + (unsigned int)objects[PISTOLS_ANIM].frame_base);
-	lara.right_arm.frame_base = (short*)((char *)lara.right_arm.frame_base + (unsigned int)objects[PISTOLS_ANIM].frame_base);
-	lara.GeneralPtr = (char *)lara.GeneralPtr + (unsigned int)malloc_buffer;
+	lara.left_arm.frame_base = (short*)((char *)lara.left_arm.frame_base + (ptrdiff_t)objects[PISTOLS_ANIM].frame_base);
+	lara.right_arm.frame_base = (short*)((char *)lara.right_arm.frame_base + (ptrdiff_t)objects[PISTOLS_ANIM].frame_base);
+	lara.GeneralPtr = (char *)lara.GeneralPtr + (ptrdiff_t)malloc_buffer;
 
 	if (lara.weapon_item == -1)
 	{
